LayoutTests/http/tests/inspector/compiler-source-mapping.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="debugger-test.js"></script>

<script>

function test()
{
    function checkMapping(compiledLineNumber, compiledColumnNumber, sourceURL, sourceLineNumber, sourceColumnNumber, mapping)
    {
        var sourceLocation = mapping.compiledLocationToSourceLocation(compiledLineNumber, compiledColumnNumber);
        InspectorTest.assertEquals(sourceURL, sourceLocation.sourceURL);
        InspectorTest.assertEquals(sourceLineNumber, sourceLocation.lineNumber);
        InspectorTest.assertEquals(sourceColumnNumber, sourceLocation.columnNumber);
    }

    function checkReverseMapping(compiledLineNumber, compiledColumnNumber, sourceURL, sourceLineNumber, mapping)
    {
        var compiledLocation = mapping.sourceLocationToCompiledLocation(sourceURL, sourceLineNumber);
        InspectorTest.assertEquals(compiledLineNumber, compiledLocation[0]);
        InspectorTest.assertEquals(compiledColumnNumber, compiledLocation[1]);
    }

    InspectorTest.runTestSuite([
        function testSimpleMapping(next)
        {
            /*
                example.js:
                0         1         2         3
                012345678901234567890123456789012345
                function add(variable_x, variable_y)
                {
                    return variable_x + variable_y;
                }

                var global = "foo";
                ----------------------------------------
                example-compiled.js:
                0         1         2         3
                012345678901234567890123456789012345
                function add(a,b){return a+b}var global="foo";
            */
            var mappingPayload = {
                "mappings":"AAASA,QAAAA,IAAG,CAACC,CAAD,CAAaC,CAAb,CACZ,CACI,MAAOD,EAAP,CAAoBC,CADxB,CAIA,IAAIC,OAAS;",
                "sources":["example.js"]
            };
            var mapping = new WebInspector.ClosureCompilerSourceMapping();
            mapping._parseMappingPayload(mappingPayload);

            checkMapping(0, 9, "example.js", 0, 9, mapping);
            checkMapping(0, 13, "example.js", 0, 13, mapping);
            checkMapping(0, 15, "example.js", 0, 25, mapping);
            checkMapping(0, 18, "example.js", 2, 4, mapping);
            checkMapping(0, 25, "example.js", 2, 11, mapping);
            checkMapping(0, 27, "example.js", 2, 24, mapping);

            checkReverseMapping(0, 0, "example.js", 0, mapping);
            checkReverseMapping(0, 17, "example.js", 1, mapping);
            checkReverseMapping(0, 18, "example.js", 2, mapping);
            checkReverseMapping(0, 29, "example.js", 4, mapping);
            checkReverseMapping(0, 29, "example.js", 5, mapping);
            InspectorTest.assertTrue(!mapping.sourceLocationToCompiledLocation("example.js", 6));

            next();
        },

        function testNoMappingEntry(next)
        {
            var mappingPayload = {
                "mappings":"AAAA,C,CAAE;",
                "sources":["example.js"]
            };
            var mapping = new WebInspector.ClosureCompilerSourceMapping();
            mapping._parseMappingPayload(mappingPayload);
            checkMapping(0, 0, "example.js", 0, 0, mapping);
            checkMapping(0, 2, "example.js", 0, 2, mapping);
            next();
        },

        function testEmptyLine(next)
        {
            var mappingPayload = {
                "mappings":"AAAA;;;CACA",
                "sources":["example.js"]
            };
            var mapping = new WebInspector.ClosureCompilerSourceMapping();
            mapping._parseMappingPayload(mappingPayload);
            checkMapping(0, 0, "example.js", 0, 0, mapping);
            checkReverseMapping(3, 1, "example.js", 1, mapping);
            next();
        },

        function testSections(next)
        {
            var mappingPayload = {
                "sections": [{
                    "offset": { "line": 0, "column": 0 },
                    "map": {
                        "mappings":"AAAA,CAEC",
                        "sources":["source1.js", "source2.js"]
                    }
                }, {
                    "offset": { "line": 2, "column": 10 },
                    "map": {
                        "mappings":"AAAA,CAEC",
                        "sources":["source2.js"]
                    }
                }
            ]};
            var mapping = new WebInspector.ClosureCompilerSourceMapping();
            mapping._parseMappingPayload(mappingPayload);
            InspectorTest.assertEquals(2, mapping.sources().length);
            checkMapping(0, 0, "source1.js", 0, 0, mapping);
            checkMapping(0, 1, "source1.js", 2, 1, mapping);
            checkMapping(2, 10, "source2.js", 0, 0, mapping);
            checkMapping(2, 11, "source2.js", 2, 1, mapping);
            next();
        },

        function testResolveSourceMappingURL(next)
        {
            var func = WebInspector.ClosureCompilerSourceMapping.prototype._canonicalizeURL;
            InspectorTest.assertEquals("http://example.com/map.json", func("http://example.com/map.json", "http://example.com/script.js"));
            InspectorTest.assertEquals("http://example.com/map.json", func("/map.json", "http://example.com/script.js"));
            InspectorTest.assertEquals("http://example.com/scripts/../maps/map.json", func("../maps/map.json", "http://example.com/scripts/script.js"));
            next();
        },

        function testCompilerScriptMapping(next)
        {
            var mapping = new WebInspector.CompilerScriptMapping();
            var script = InspectorTest.createScriptMock("compiled.js", 0, 0, true, "");
            script.sourceMapURL = "http://localhost:8000/inspector/resources/source-map.json";
            mapping.addScript(script);

            var uiSourceCodeList = mapping.uiSourceCodeList();
            InspectorTest.assertEquals(2, uiSourceCodeList.length);
            InspectorTest.assertEquals("http://localhost:8000/inspector/resources/source1.js", uiSourceCodeList[0].url);
            InspectorTest.assertEquals("http://localhost:8000/inspector/resources/source2.js", uiSourceCodeList[1].url);

            InspectorTest.checkUILocation(uiSourceCodeList[0], 4, 4, mapping.rawLocationToUILocation(WebInspector.debuggerModel.createRawLocation(script, 0, 81)));
            InspectorTest.checkUILocation(uiSourceCodeList[0], 5, 4, mapping.rawLocationToUILocation(WebInspector.debuggerModel.createRawLocation(script, 0, 93)));
            InspectorTest.checkUILocation(uiSourceCodeList[1], 7, 4, mapping.rawLocationToUILocation(WebInspector.debuggerModel.createRawLocation(script, 1, 151)));

            InspectorTest.checkRawLocation(script, 0, 42, mapping.uiLocationToRawLocation(uiSourceCodeList[0], 3, 10));
            InspectorTest.checkRawLocation(script, 1, 85, mapping.uiLocationToRawLocation(uiSourceCodeList[1], 0, 0));
            InspectorTest.checkRawLocation(script, 1, 110, mapping.uiLocationToRawLocation(uiSourceCodeList[1], 5, 2));

            uiSourceCodeList[0].requestContent(didRequestContent1);

            function didRequestContent1(mimeType, content)
            {
                InspectorTest.assertEquals(0, content.indexOf("window.addEventListener"));
                uiSourceCodeList[1].requestContent(didRequestContent2);
            }

            function didRequestContent2(mimeType, content)
            {
                InspectorTest.assertEquals(0, content.indexOf("function ClickHandler()"));
                next();
            }
        },

        function testCompilerSourceMappingCouldNotBeLoaded(next)
        {
            var mainScriptMapping = new WebInspector.MainScriptMapping();
            mainScriptMapping.addEventListener(WebInspector.MainScriptMapping.Events.UISourceCodeListChanged, uiSourceCodeListChanged);

            var script = InspectorTest.createScriptMock("compiled.js", 0, 0, true, "");
            script.sourceMapURL = "http://localhost:8000/inspector/resources/source-map.json_";
            mainScriptMapping.addScript(script);

            function uiSourceCodeListChanged(event)
            {
                InspectorTest.assertEquals(1, event.data.addedItems.length);
                InspectorTest.assertEquals("compiled.js", event.data.addedItems[0].url);
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests ClosureCompilerSourceMapping.</p>
</body>
</html>
